narrowing visibility for the benchmark module
