+++ platform-sdk/swirlds-logging-benchmark/src/main/java/com/swirlds/logging/benchmark/ConfigureSwirldsLog.java
@@ -22,11 +22,15 @@ import com.swirlds.logging.api.extensions.handler.LogHandler;
 import com.swirlds.logging.api.internal.LoggingSystem;
 import com.swirlds.logging.api.internal.configuration.ConfigLevelConverter;
 import com.swirlds.logging.api.internal.configuration.MarkerStateConverter;
+import com.swirlds.logging.benchmark.config.LoggingHandlingType;
+import com.swirlds.logging.benchmark.config.LoggingImplementation;
+import com.swirlds.logging.benchmark.util.LogFileUtlis;
 import com.swirlds.logging.external.benchmark.BenchmarkFactory;
+import edu.umd.cs.findbugs.annotations.NonNull;
 
 public class ConfigureSwirldsLog {
 
-    public static LoggingSystem configureFileLogging() {
+    public static @NonNull LoggingSystem configureFileLogging() {
         final String logFile = LogFileUtlis.provideLogFilePath(LoggingImplementation.SWIRLDS, LoggingHandlingType.FILE);
         final Configuration configuration = ConfigurationBuilder.create()
                 .withConverter(new ConfigLevelConverter())
@@ -44,7 +48,7 @@ public class ConfigureSwirldsLog {
         return loggingSystem;
     }
 
-    public static LoggingSystem configureConsoleLogging() {
+    public static @NonNull LoggingSystem configureConsoleLogging() {
         final Configuration configuration = ConfigurationBuilder.create()
                 .withConverter(new ConfigLevelConverter())
                 .withConverter(new MarkerStateConverter())
@@ -61,7 +65,7 @@ public class ConfigureSwirldsLog {
         return loggingSystem;
     }
 
-    public static LoggingSystem configureFileAndConsoleLogging() {
+    public static @NonNull LoggingSystem configureFileAndConsoleLogging() {
         final String logFile =
                 LogFileUtlis.provideLogFilePath(LoggingImplementation.SWIRLDS, LoggingHandlingType.CONSOLE_AND_FILE);
         final Configuration configuration = ConfigurationBuilder.create()
