+++ platform-sdk/swirlds-logging-benchmark/src/main/java/com/swirlds/logging/benchmark/LogWithLog4J.java
@@ -16,6 +16,7 @@
 
 package com.swirlds.logging.benchmark;
 
+import com.swirlds.logging.benchmark.util.BenchmarkUtils;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.Marker;
